+++ src/main/scala/uni/time/ExampleGenerator.scala
@@ -2,8 +2,8 @@ package uni.time
 
 import scala.util.Random
 
-object ExampleGenerator:
-  import SmarterParse.{Shape, Pattern}
+object ExampleGeneratorSmarter:
+  import SmartParse.{Shape, Pattern}
 
   private var rng: Random = new Random(0)
   def seed(n: Int): Unit =
@@ -11,7 +11,6 @@ object ExampleGenerator:
 
   // Delimiters for numeric patterns
   private val delimiters = Array("/", "-", ".", " ")
-
   def exampleForPattern(pattern: String): (String, Shape) =
     var ex = ""
     var inferred: Shape = Shape.Unknown
@@ -19,8 +18,8 @@ object ExampleGenerator:
     // Keep generating until classifier accepts it
     while inferred == Shape.Unknown do
       ex = generateRawExample(pattern)
-      inferred = SmarterParse.classify(ex)
-      if uni.verboseUni && inferred == Shape.Unknown then
+      inferred = SmartParse.classify(ex)
+      if (uni.verboseUni && inferred == Shape.Unknown) then
         System.err.printf(s"rejected pattern: [$ex]\n")
         //sys.error(s"rejected pattern [$ex]")
 
@@ -28,12 +27,15 @@ object ExampleGenerator:
 
   private def generateRawExample(pattern: String): String =
     pattern match
-      case Pattern.YMD.pattern          => expandYMD()
-      case Pattern.MDY.pattern          => expandMDY()
-      case Pattern.DMY.pattern          => expandDMY()
-      case Pattern.MonthDayYear.pattern => expandMonthDayYear()
-      case Pattern.ISO8601Strict.pattern => expandISO8601Strict()
-      case other =>
+      case Pattern.YMD.pattern                  => expandYMD()
+      case Pattern.MDY.pattern                  => expandMDY()
+      case Pattern.DMY.pattern                  => expandDMY()
+      case Pattern.MonthDayYear.pattern         => expandMonthDayYear()
+      case Pattern.DayMonthYear.pattern         => expandDayMonthYear()
+      case Pattern.MonthDayYearWithTime.pattern => expandMonthDayYearWithTime()
+      case Pattern.MDYWithTime.pattern          => expandMDYWithTime()
+      case Pattern.ISO8601Strict.pattern        => expandISO8601Strict()
+      case other           =>
         throw new RuntimeException(s"Unknown pattern: $other")
 
   // -----------------------
@@ -46,23 +48,40 @@ object ExampleGenerator:
   private def expandYMD(): String =
     val y = 2000 + rng.nextInt(30)
     val m = 1 + rng.nextInt(12)
-    val d = 1 + rng.nextInt(28)
+    val maxDay = maxDays(m-1)
+    val d = 1 + rng.nextInt(maxDay)
     val d1 = delim
     s"$y$d1$m$d1$d"
 
   private def expandMDY(): String =
     val m = 1 + rng.nextInt(12)
-    val d = 1 + rng.nextInt(28)
+    val maxDay = maxDays(m-1)
+    val d = 1 + rng.nextInt(maxDay)
     val y = 2000 + rng.nextInt(30)
     val d1 = delim
     s"$m$d1$d$d1$y"
 
-  private def expandDMY(): String =
-    val d = 1 + rng.nextInt(28)
+  private def expandMDYWithTime(): String = 
     val m = 1 + rng.nextInt(12)
+    val maxDay = maxDays(m-1)
+    val d = 1 + rng.nextInt(maxDay)
+    val y = 2000 + rng.nextInt(30)
+    val d1 = delim
+    val hh = rng.nextInt(24)
+    val mm = rng.nextInt(60)
+    val ss = rng.nextInt(60)
+    f"$m%02d$d1$d%02d $hh%02d:$mm%02d:$ss%02d $y"
+
+  lazy val maxDays = IndexedSeq(31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
+
+  private def expandDMY(): String = {
+    val m = 1 + rng.nextInt(12)
+    val maxDay = maxDays(m - 1)
+    val d = 13 + rng.nextInt(maxDay - 12)
     val y = 2000 + rng.nextInt(30)
     val d1 = delim
     s"$d$d1$m$d1$y"
+  }
 
   private def expandMonthDayYear(): String =
     val months = Array("Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec")
@@ -76,6 +95,33 @@ object ExampleGenerator:
     else
       val d1 = delim
       s"$month$d1$d$d1$y"
+
+  private def expandDayMonthYear(): String =
+    val months = Array("Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec")
+    val month = months(rng.nextInt(12))
+    val d = 1 + rng.nextInt(28)
+    val y = 2000 + rng.nextInt(30)
+
+    // 80% classic "Aug 11, 2020", 20% delimiter variants
+    if rng.nextDouble() < 0.8 then
+      s"$d $month $y"
+    else
+      val d1 = delim
+      s"$d $month$d1$y"
+
+  private def expandMonthDayYearWithTime(): String =
+    val months = Array("Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec")
+    val month = months(rng.nextInt(12))
+    val d = 1 + rng.nextInt(28)
+    val y = 2000 + rng.nextInt(30)
+
+    val hh = rng.nextInt(24)
+    val mm = rng.nextInt(60)
+    val ss = rng.nextInt(60)
+    val time = f"$hh%02d:$mm%02d:$ss%02d"
+
+    val d1 = delim
+    s"$month$d1$d $time $d1$y"
 
   private def expandISO8601Strict(): String =
     val y = 2000 + rng.nextInt(30)
